Add layouts to SignUpPage. Add console.logs for debugging.
